-Fix several incorrect variable types -Change ref parameters to out parameters that do not explicitly specify [In, Out] -Fix IntPtr buffers that should be [In, Out] ref

+++ ManagedCorDebug/Native/Interfaces/SOS/ISOSDacInterface.cs
@@ -401,7 +401,7 @@ namespace ManagedCorDebug
 
         [PreserveSig]
         HRESULT GetDacModuleHandle(
-            out long phModule);
+            out IntPtr phModule);
 
         [PreserveSig]
         HRESULT GetRCWData(
